An attempt to fix incorrect view size on scaled resolutions

+++ src/game/renderer.cpp
@@ -48,11 +48,11 @@
 
 Renderer * Renderer::m_instance = nullptr;
 
-Renderer::Renderer(int hRes, int vRes, int fullHRes, int fullVRes, int pixelScale, bool fullScreen, MCGLScene & glScene)
-  : m_hRes { hRes * pixelScale }
-  , m_vRes { vRes * pixelScale }
-  , m_fullHRes { fullHRes * pixelScale }
-  , m_fullVRes { fullVRes * pixelScale }
+Renderer::Renderer(int hRes, int vRes, int fullHRes, int fullVRes, double pixelScale, bool fullScreen, MCGLScene & glScene)
+  : m_hRes { static_cast<int>(hRes * pixelScale) }
+  , m_vRes { static_cast<int>(vRes * pixelScale) }
+  , m_fullHRes { static_cast<int>(fullHRes * pixelScale) }
+  , m_fullVRes { static_cast<int>(fullVRes * pixelScale) }
   , m_pixelScale { pixelScale }
   , m_fullScreen { fullScreen }
   , m_glScene { glScene }
@@ -79,15 +79,15 @@ void Renderer::initialize()
     if (!m_fullScreen)
     {
         // Set window size & disable resize
-        const int windowHRes = m_hRes / m_pixelScale;
-        const int windowVRes = m_vRes / m_pixelScale;
+        const int windowHRes = static_cast<int>(m_hRes / m_pixelScale);
+        const int windowVRes = static_cast<int>(m_vRes / m_pixelScale);
         resize(windowHRes, windowVRes);
         setMinimumSize(QSize(windowHRes, windowVRes));
         setMaximumSize(QSize(windowHRes, windowVRes));
 
         // Try to center the window
-        const int screenHRes = m_fullHRes / m_pixelScale;
-        const int screenVRes = m_fullVRes / m_pixelScale;
+        const int screenHRes = static_cast<int>(m_fullHRes / m_pixelScale);
+        const int screenVRes = static_cast<int>(m_fullVRes / m_pixelScale);
         setPosition(screenHRes / 2 - windowHRes / 2, screenVRes / 2 - windowHRes / 2);
     }
